change testing to fastmcp
